# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/b/auth/staff/login": {
      "post": {
        "operationId": "B_AuthController_staffLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns access token and staff profile"
          },
          "401": {
            "description": "Invalid credentials or inactive account"
          }
        },
        "summary": "Log in as Staff (Admin/Editor) and receive a JWT access token",
        "tags": [
          "[B] auth staff"
        ]
      }
    },
    "/api/b/auth/staff/register": {
      "post": {
        "operationId": "B_AuthController_staffRegister",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffRegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Register successful, returns staff profile"
          },
          "401": {
            "description": "Not logged in or unauthorized"
          },
          "403": {
            "description": "Logged in but not permitted (non-admin)"
          },
          "409": {
            "description": "A staff with the credential exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a new staff member (admin only)",
        "tags": [
          "[B] auth staff"
        ]
      }
    },
    "/api/b/auth/staff/forgot_password": {
      "post": {
        "operationId": "B_AuthController_staffForgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request accepted; a reset code is sent to the email if it is registered"
          }
        },
        "summary": "Get reset code if password is forgotten",
        "tags": [
          "[B] auth staff"
        ]
      }
    },
    "/api/b/auth/staff/reset_password": {
      "post": {
        "operationId": "B_AuthController_staffResetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset password successful"
          },
          "400": {
            "description": "Invalid or expired reset code"
          }
        },
        "summary": "Reset password with reset code",
        "tags": [
          "[B] auth staff"
        ]
      }
    },
...
  },
  "info": {
    "title": "Blog Management API",
    "description": "API for blog management system with user authentication and post management",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "StaffLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@email.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "StaffRegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@email.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "role": {
            "type": "number",
            "example": 1
          },
          "status": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "role",
          "status"
        ]
      },
      "StaffForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@email.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "StaffResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@email.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "resetCode": {
            "type": "string",
            "example": "123789"
          }
        },
        "required": [
          "email",
          "password",
          "resetCode"
        ]
      },
...
    }
  }
}